refactored UI for account setup
