+++ src/components/account-setup/AccountSetup.js
@@ -1,6 +1,7 @@
 import React, { Component } from 'react';
+
 import Account from './Account';
-import Leaders from './Leaders';
+import LeadersContacts from './LeadersContacts';
 
 const CLASS_NAMES = {
     ACCOUNT_SETUP: 'account-setup',
@@ -14,12 +15,8 @@ class AccountSetup extends Component {
             <div className={CLASS_NAMES.ACCOUNT_SETUP}>
                 <Account />
                 <div className={CLASS_NAMES.LEADERS_AND_CONTACTS}>
-                    <Leaders title={'Leaders'}
-                        name={'Julia Libder'}
-                        phoneNumber={'+46 591 409173'}
-                        email={'julia.libder@amco.com'}
-                        actionName={'Add Leader'} />
-                    <Leaders title={'VATBox Contacts'}
+                    <LeadersContacts title={'Leaders'} />
+                    <LeadersContacts title={'VATBox Contacts'}
                         name={'Yuval Armoni'}
                         phoneNumber={'+46 591 409173'}
                         email={'yuval.armoni@amco.com'}
